# SharpAI.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SharpAI.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Audio/audios": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "includeWaveforms",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "width",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 512
            }
          },
          {
            "name": "height",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudioObjInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudioObjInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudioObjInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Audio/audios/data": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sampleRate",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "channels",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "bitDepth",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Audio/audios/waveform": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "width",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 512
            }
          },
          {
            "name": "height",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 64
            }
          },
          {
            "name": "samplesPerPixel",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "backColorHex",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "#FFFFFF"
            }
          },
          {
            "name": "graphColorHex",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "#000000"
            }
          },
          {
            "name": "useCache",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "maxWorkers",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Audio/audios/upload": {
      "post": {
        "tags": [
          "Audio"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjInfo"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Audio/audios/record": {
      "post": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "sampleRate",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 16000
            }
          },
          {
            "name": "channels",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "bitDepth",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjInfo"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Audio/audios/resample": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sampleRate",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 16000
            }
          },
          {
            "name": "channels",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "bitDepth",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Audio/audios/rechannel": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sampleRate",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 16000
            }
          },
          {
            "name": "channels",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "bitDepth",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioObjData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Audio/audios/export": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exportDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bits",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Audio/audios/download": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bits",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Audio/download": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bits",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Audio/audios/delete": {
      "delete": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioObjData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sampleRate": {
            "type": "integer",
            "format": "int32"
          },
          "channels": {
            "type": "integer",
            "format": "int32"
          },
          "bitDepth": {
            "type": "integer",
            "format": "int32"
          },
          "durationInSeconds": {
            "type": "number",
            "format": "double"
          },
          "data": {
            "type": "string",
            "nullable": true
          },
          "length": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "AudioObjInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "filePath": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "durationInSeconds": {
            "type": "number",
            "format": "double"
          },
          "waveformPreview": {
            "$ref": "#/components/schemas/ImageObjData"
          }
        },
        "additionalProperties": false
      },
      "ImageObjData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}